Feat/growingsetup service

+++ src/main/java/dk/via/group1/urbanmicrofarm_backend/database/repository/GrowingSetupRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 
-public interface GrowingSetupRepository extends JpaRepository<GrowingSetupEntity, Long> {
+public interface GrowingSetupRepository extends JpaRepository<GrowingSetupEntity, Integer> {
 
   List<GrowingSetupEntity> findByEmail(String email);
 }
