+++ src/main/java/dk/via/group1/urbanmicrofarm_backend/mapper/dbMapper/GrowingSetupPersistenceMapper.java
@@ -17,7 +17,7 @@ public class GrowingSetupPersistenceMapper {
 
   public GrowingSetup toDomain(GrowingSetupEntity entity) {
     return new GrowingSetup(
-        entity.getSetupId().intValue(),
+        entity.getSetupId(),
         entity.getSerialNumber(),
         entity.getLocation(),
         null
